Infer transaction currency from API
