+++ src/assets/cosmos/provider/bigdipper/README.md
@@ -7,3 +7,9 @@ https://bigdipper.live/
 ## API
 
 https://gql.akash.forbole.com/v1/graphql
+
+## All Apis
+https://github.com/cosmos/chain-registry/blob/master/mars/chain.json
+
+# denom_trace
+https://v1.cosmos.network/rpc/ibc
